10039: Move temp scripts into run-once-scripts and move ci scripts into circleci folder
