+++ .circleci/config.yml
@@ -372,7 +372,7 @@ jobs:
             set -e
       - run:
           name: Disable Test User
-          command: npx ts-node --transpile-only ./scripts/disable-smoketest-user.js
+          command: npx ts-node --transpile-only ./scripts/circleci/disable-smoketest-user.js
       - run:
           name: Kill Build if Smoketests Failed
           command: ./scripts/kill-circle-build.sh
@@ -560,7 +560,7 @@ jobs:
       - run:
           name: Setup the deploy table for a glue job
           command: |
-            ./scripts/setup-deploy-table-for-glue-job.sh
+            ./scripts/circleci/setup-deploy-table-for-glue-job.sh
       - run:
           name: Delete all existing dynamodb tables and elasticsearch clusters
           command: |
@@ -742,7 +742,7 @@ jobs:
       - run:
           name: Sync S3 buckets with timeout
           command: |
-            ./scripts/sync-s3-buckets-with-timeout.sh 175m
+            ./scripts/circleci/sync-s3-buckets-with-timeout.sh 175m
 
   ######################## GLUE JOB: Jobs that run in prod ########################
