+++ package.json
@@ -236,10 +236,10 @@
     "test:shared:ci": "npm run build:assets && jest --config shared/jest-shared.config.ts --coverageDirectory=./coverage/${NODE_INDEX} ${TESTFILES}",
     "test:watch": "jest --watch .*test\\.js$",
     "test": "echo 'Best run with start:api:ci and start:client:ci' && sleep 3 && CI=true npx run-s test:shared test:api test:client test:pa11y cypress lint --prefix=web-client/pa11y/",
-    "verify-authorizers": "./scripts/verify-authorizers.sh",
-    "verify-private-elasticsearch": "./scripts/verify-private-elasticsearch.sh",
-    "verify-private-s3-buckets": "./scripts/verify-private-s3-buckets.sh",
-    "verify-cognito-lambda-triggers": "./scripts/verify-cognito-lambda-triggers.sh",
+    "verify-authorizers": "./scripts/circleci/verify-authorizers.sh",
+    "verify-private-elasticsearch": "./scripts/circleci/verify-private-elasticsearch.sh",
+    "verify-private-s3-buckets": "./scripts/circleci/verify-private-s3-buckets.sh",
+    "verify-cognito-lambda-triggers": "./scripts/circleci/verify-cognito-lambda-triggers.sh",
     "visualize-bundle": "esbuild-visualizer --metadata ./metadata.json && open stats.html",
     "watch": "nodemon -e yml --exec sls offline start"
   },
