+++ scripts/circleci/setup-deploy-table-for-glue-job.sh
