+++ scripts/circleci/verify-authorizers.sh
