+++ scripts/circleci/verify-cognito-lambda-triggers.sh
