+++ scripts/circleci/verify-private-elasticsearch.sh
