+++ scripts/circleci/verify-private-s3-buckets.sh
