+++ scripts/run-once-scripts/change-chief-judge.sh
