+++ scripts/run-once-scripts/create-many-cases-with-electronic-and-paper-service.sh
