+++ scripts/run-once-scripts/delete-old-infrastructure-pre-migration.sh
